refactror and add abstracts
