+++ src/clients/artikkelit/utils.js
@@ -1,58 +1,3 @@
-import {idbSet, idbKeys} from "/artikkelit/indexDB.js"
-
-export function addTempOrganization({organizationName, code, organizationShortTerm, note}) {
-  const oldItems = window.sessionStorage.getItem('melinda-temp-organizations');
-  if (oldItems === null) {
-    window.sessionStorage.setItem('melinda-temp-organizations', JSON.stringify([{organizationName, code, organizationShortTerm, note}]));
-    return true;
-  }
-
-  const oldItemsArray = JSON.parse(oldItems);
-  const hasValue = oldItemsArray.some(item => item.code === `${code}`);
-
-  if (hasValue) {
-    return false;
-  }
-
-  if (!code && organizationShortTerm) {
-    window.sessionStorage.setItem('melinda-temp-organizations', JSON.stringify([...JSON.parse(oldItems), {organizationName, code: organizationShortTerm, organizationShortTerm, note}]));
-  }
-
-  window.sessionStorage.setItem('melinda-temp-organizations', JSON.stringify([...JSON.parse(oldItems), {organizationName, code, organizationShortTerm, note}]));
-  return true;
-}
-
-export function removeTempOrganization(code) {
-  const oldItems = window.sessionStorage.getItem('melinda-temp-organizations');
-  if (oldItems === null) {
-    return true;
-  }
-
-  const oldItemsArray = JSON.parse(oldItems);
-  const hasValue = oldItemsArray.some(item => item.code === `${code}`);
-
-  if (hasValue) {
-    const newItems = oldItemsArray.filter(item => item.code !== `${code}`);
-    window.sessionStorage.setItem('melinda-temp-organizations', JSON.stringify(newItems));
-    return true;
-  }
-
-  return true;
-}
-
-export function resetTempOrganizationList() {
-  window.sessionStorage.removeItem('melinda-temp-organizations');
-}
-
-export function getTempOrganizationList() {
-  const oldItems = window.sessionStorage.getItem('melinda-temp-organizations');
-  if (oldItems === null) {
-    return [];
-  }
-
-  return JSON.parse(oldItems);
-}
-
 export function createP(value, before = '', after = '') {
   const p = document.createElement('p');
   p.innerHTML = `${before}${value}${after}`;
